requirements.txt
pip
setuptools
wheel
django
gunicorn
uwsgi
django-extensions
pygraphviz
pydotplus
django-trackstats
django-taggit
django-allauth
django-filter
django-filters
psycopg2-binary
django-bootstrap-form
pygithub
markdown2
coverage
coveralls
beautifulsoup4
whitenoise
